src/core/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
alembic>=1.13.0
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
psycopg2-binary>=2.9.9
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Dev dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.26.0
ruff>=0.1.0
